Hacky but working validation of moves
All Validatable objects have a dictionary of rules, that tie a
specific function (game action) to a specific rule. This is used
via the `checkRule()` decorator, which is used to validate card
movements.

Validation for discarding from hand onto discard player and for
using the replacement piles seems to be working.

Updated DSL defintion of speed to include that ACE can be played
on KING, and KING can be played on ACE.