Got in-fix Bartok working
The issue was that references to previously defined rules act as
lambdas; for the shunting yard algo. to recognize these lambdas, they
need to have the "Fn." prefixed in their invocations.

Added precedence data to the global environment for built-in operators

Cleaned up the shunting yard code (in `build_list`)