Speed appears to be work, but it's ugly
The rules can't be broken down b/c of the way the stack machine works
currently. For example, if the validDiscard move was expressed as
validDiscardPlusMinusOne or validDiscardWrapAround, it wouldn't work correctly
b/c after running the first function (left of or symbol), it'll have a boolean
value (the result of the function) in the list, and when starting to run the
second function, it'll erroneously assume the boolean value to be an argument
to the second function.

One way to fix this is to check whether the value that is to be passed as an
argument to any function is of the right type (move or card). This will fix
the issue, but seems to cause a problem with the validReplacement rule.

Also, expressions like not(a) and not(b) apparently don't work. Soln is to
apply DeMorgans to get not(a or b) which will work.